move appresolve package to apps/common
