+++ pkg/cli/commands/apps/files/show.go
@@ -8,7 +8,7 @@ import (
 	"strings"
 	"time"
 
-	resolve "github.com/superplanehq/superplane/pkg/cli/appresolve"
+	"github.com/superplanehq/superplane/pkg/cli/commands/apps/common"
 	"github.com/superplanehq/superplane/pkg/cli/core"
 )
 
@@ -20,7 +20,7 @@ func (c *ShowCommand) Execute(ctx core.CommandContext) error {
 		return err
 	}
 
-	canvasID, err := resolve.ResolveAppNameOrIDArg(ctx, canvasTarget)
+	canvasID, err := common.ResolveAppNameOrIDArg(ctx, canvasTarget)
 	if err != nil {
 		return err
 	}
